Made some structure
