+++ src/assets/tasks_math_medium.json
@@ -12,4 +12,3 @@
       "answer": "2"
     }
   ]
-
